+++ api/src/main/kotlin/com/crowdproj/marketplace/product/groups/api/ApiMapper.kt
@@ -1,4 +1,4 @@
-package apimappers
+package com.crowdproj.marketplace.product.groups.api
 import com.fasterxml.jackson.databind.ObjectMapper
 
 val apiMapper = ObjectMapper().apply {}
