+++ common/src/commonMain/kotlin/com.crowdproj.marketplace.product.groups.common/Constants.kt
@@ -1,3 +1,5 @@
+package com.crowdproj.marketplace.product.groups.common
+
 import kotlinx.datetime.Instant
 
 private val INSTANT_NONE = Instant.fromEpochMilliseconds(Long.MIN_VALUE)
