+++ common/src/commonMain/kotlin/com.crowdproj.marketplace.product.groups.common/models/ProductGroupCommand.kt
@@ -1,4 +1,4 @@
-package models
+package com.crowdproj.marketplace.product.groups.common.models
 
 enum class ProductGroupCommand {
     NONE,
