+++ common/src/commonMain/kotlin/com.crowdproj.marketplace.product.groups.common/models/ProductGroupFilter.kt
@@ -1,4 +1,4 @@
-package models
+package com.crowdproj.marketplace.product.groups.common.models
 
 data class ProductGroupFilter(
     var id: ProductGroupId = ProductGroupId.NONE,
