+++ mappers/src/main/kotlin/com/crowdproj/marketplace/product/groups/mappers/exceptions/UnknownRequestClass.kt
@@ -1,4 +1,4 @@
-package exceptions
+package com.crowdproj.marketplace.product.groups.mappers.exceptions
 
 class UnknownRequestClass(clazz: Class<*>) : RuntimeException("Class $clazz cannot be mapped to MkplContext") {
 }
