--- mappers/src/main/kotlin/exceptions/UnknownProductGroupCommand.kt
@@ -1,5 +1,0 @@
-package exceptions
-
-import models.ProductGroupCommand
-
-class UnknownProductGroupCommand(command: ProductGroupCommand) : Throwable("Wrong command $command at mapping toTransport stage")
